move config file
