+++ examples/node/Examples/create-bundled-realm-rn.js
@@ -1,5 +1,5 @@
 import { Realm } from "realm";
-import { REALM_APP_ID } from "./config.js";
+import { REALM_APP_ID } from "../config.js";
 
 const Dog = {
   name: "Dog",
