+++ examples/node/config.js
